Check modifications in contracts before add/upgrade proxy
This PR adds to the network package.zos file the bytecode of the
contract at the time it was deployed. It is then compared to the
bytecode of the contract locally before creating or upgrading a
proxy. If there was a change, then the command fails; this can be
overridden via a --force flag to either command.